Create a carousel section for jobs
This is pay to win! To have a better chance of securing a specific
niche job, let alone any job at all, the viewer can check out the
different plans to see if it interests them.

This is beneficial for them to increase their chances of finding a job
and it means more profit. Because we work like Tinder. Pay to win.

Ha ha!